+++ src/api_traits/telegram_api.rs
@@ -9,8 +9,6 @@ use crate::api_params::AnswerWebAppQueryParams;
 use crate::api_params::ApproveChatJoinRequestParams;
 use crate::api_params::BanChatMemberParams;
 use crate::api_params::BanChatSenderChatParams;
-use crate::api_params::BotDescription;
-use crate::api_params::BotShortDescription;
 use crate::api_params::CloseForumTopicParams;
 use crate::api_params::CloseGeneralForumTopicParams;
 use crate::api_params::CopyMessageParams;
@@ -54,7 +52,6 @@ use crate::api_params::GetUpdatesParams;
 use crate::api_params::GetUserProfilePhotosParams;
 use crate::api_params::HideGeneralForumTopicParams;
 use crate::api_params::InputMedia;
-use crate::api_params::InputSticker;
 use crate::api_params::LeaveChatParams;
 use crate::api_params::Media;
 use crate::api_params::PinChatMessageParams;
@@ -110,6 +107,8 @@ use crate::api_params::UnpinAllForumTopicMessagesParams;
 use crate::api_params::UnpinChatMessageParams;
 use crate::api_params::UploadStickerFileParams;
 use crate::objects::BotCommand;
+use crate::objects::BotDescription;
+use crate::objects::BotShortDescription;
 use crate::objects::Chat;
 use crate::objects::ChatAdministratorRights;
 use crate::objects::ChatInviteLink;
@@ -117,6 +116,7 @@ use crate::objects::ChatMember;
 use crate::objects::File as FileObject;
 use crate::objects::ForumTopic;
 use crate::objects::GameHighScore;
+use crate::objects::InputSticker;
 use crate::objects::MenuButton;
 use crate::objects::Message;
 use crate::objects::MessageId;
